Refactor: ProductManage 분리
