+++ src/refactoring/components/admin/ProductManage.tsx
@@ -1,7 +1,6 @@
-import { useState } from "react";
 import { Product } from "../../../types";
-import { initialNewProducts } from "../../constants";
 import ProductList from "./ProductList";
+import ProductAdd from "./ProductAdd";
 
 interface Props {
   products: Product[];
@@ -10,97 +9,10 @@ interface Props {
 }
 
 const ProductManage = ({ products, onProductAdd, onProductUpdate }: Props) => {
-  const [showNewProductForm, setShowNewProductForm] = useState(false);
-  const [newProduct, setNewProduct] =
-    useState<Omit<Product, "id">>(initialNewProducts);
-
-  const handleAddNewProduct = () => {
-    //시간으로 id 설정
-    const productWithId = { ...newProduct, id: Date.now().toString() };
-    onProductAdd(productWithId); //전역적으로 새로운 상품을 추가
-    setNewProduct(initialNewProducts);
-    setShowNewProductForm(false); //새 상품 추가 후, 폼 닫기
-  };
-
   return (
     <div>
       <h2 className="text-2xl font-semibold mb-4">상품 관리</h2>
-      {/* 취소 또는 새상품 추가 */}
-      <button
-        onClick={() => setShowNewProductForm(!showNewProductForm)}
-        className="bg-green-500 text-white px-4 py-2 rounded mb-4 hover:bg-green-600"
-      >
-        {showNewProductForm ? "취소" : "새 상품 추가"}
-      </button>
-      {/* 새로운 상품을 추가하기 위한 폼 */}
-      {showNewProductForm && (
-        <div className="bg-white p-4 rounded shadow mb-4">
-          <h3 className="text-xl font-semibold mb-2">새 상품 추가</h3>
-          <div className="mb-2">
-            <label
-              htmlFor="productName"
-              className="block text-sm font-medium text-gray-700"
-            >
-              상품명
-            </label>
-            <input
-              id="productName"
-              type="text"
-              value={newProduct.name}
-              onChange={(e) =>
-                setNewProduct({ ...newProduct, name: e.target.value })
-              }
-              className="w-full p-2 border rounded"
-            />
-          </div>
-          <div className="mb-2">
-            <label
-              htmlFor="productPrice"
-              className="block text-sm font-medium text-gray-700"
-            >
-              가격
-            </label>
-            <input
-              id="productPrice"
-              type="number"
-              value={newProduct.price}
-              onChange={(e) =>
-                setNewProduct({
-                  ...newProduct,
-                  price: parseInt(e.target.value),
-                })
-              }
-              className="w-full p-2 border rounded"
-            />
-          </div>
-          <div className="mb-2">
-            <label
-              htmlFor="productStock"
-              className="block text-sm font-medium text-gray-700"
-            >
-              재고
-            </label>
-            <input
-              id="productStock"
-              type="number"
-              value={newProduct.stock}
-              onChange={(e) =>
-                setNewProduct({
-                  ...newProduct,
-                  stock: parseInt(e.target.value),
-                })
-              }
-              className="w-full p-2 border rounded"
-            />
-          </div>
-          <button
-            onClick={handleAddNewProduct}
-            className="w-full bg-blue-500 text-white p-2 rounded hover:bg-blue-600"
-          >
-            추가
-          </button>
-        </div>
-      )}
+      <ProductAdd onProductAdd={onProductAdd} />
       <ProductList products={products} onProductUpdate={onProductUpdate} />
     </div>
   );
